proxy: Prepend data by its length in the client<->proxy protocol
I realized that my assumption sendmsg() would create a barrier on the
read side was wrong. If one writes data to one end of the connection:

  - write([32 bytes of data])
  - sendmsg([1 byte of data], [fd])

I was assuming the first read() wouldn't squash the 1 byte sent with
sendmsg() with the write 32 bytes.

Because this isn't a guarantee, we need to have a way to either:
  - know an fd has been attached (through a set flag in a header) so we
    can use a recvmsg
  - know how much data we need to read at all time so we can limit the
    reads to the right amount of data

I chose 2/ because it fits more with what I had, but the protocol might
still change, who knows.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>