proxy: Close the connection when failing to send a fd
The behaviour for the other error cases is well defined: we close the
client connection on error. However the case where we fail to send a fd
seems to have been forgotten.

To try and minimize the chances of this happening again, make
protocol.server() return an error and there's only one site where we,
then, close the connection.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>