proxy: Create runtime directory if needed when not started by systemd
It's expected that cc-proxy is socket activated by systemd in
production, but, the proxy also has a path where it creates the socket
itself. This is useful for debugging as it allows people to launch the
proxy from the command line directly.

Unfortunately, we were not able to create the socket in that second
path when the /run/cc-oci-runtime/ directory wasn't already created
(systemd will do that for us in the socket activated case).

Fix this by creating the /run/cc-oci-runtime directory if needed.

Fixes: https://github.com/01org/cc-oci-runtime/issues/459
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>