proxy/api: Add the path to the VM console in the Hello payload
It can be handy to have the runtime give us the path to the console
socket as can instruct the VM kernel or processes inside the VM to
output debug logs to the console.

We do have one process that stays around for the whole lifetime of the
VM, cc-proxy. It's a good place to extract console data, so we make the
runtime tell cc-proxy where that the console socket is.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>