proxy: Close VM object when detecting the end of the VM process
There are cases when Bye isn't sent to the proxy: when called by docker kill
with the KILL signal.

Instead, we can do the VM object cleanup (ie. close the opened sockets to the
qemu process) when we detect the VM process is gone.

We still keep the Bye payload, it's the one in charge of making the VM object
disapear from the API. This payload could also be useful in the future if we
want to implement some robustness measure that checks the qemu process is
indeed gone after a Bye (and do something about it if not).

Fixes: https://github.com/01org/cc-oci-runtime/issues/542

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>