proxy: Reorder response and fd
In:

  commit 50d42dc434f283ccb6e5daa17ea98d376b86fc4c
  Author: Samuel Ortiz <sameo@linux.intel.com>
  Date:   Mon Dec 19 15:00:25 2016 +0100

      proxy: Simplify fd passing and the message read loop

Samuel changed the order to send the fd before the response and simplify
the reading loop.

In this commit, I put it back the other way: response, then fd. It
allows for a more natural checking of the response status before trying
to receive the fd. If the allocateIO fails, we don't even try to read a
fd from the proxy.

That change is motivated by another one: we now have a
cc_proxy_receive_fd() function that is orthogonal to reading the
response back from the proxy. That means we can use it several times in
a row to retrieve more than one fd, a useful feature we'd like to use
very soon.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>