proxy: Fix the VM lost detection site
I placed the signal that we've lost the VM when the read() on the I/O
socket was failing. That was missing the rest of hyper to client I/O
copy loop: there are other places where having lost the VM will result
in this goroutine returning.

So, instead, place the VM lost signal at the end of the goroutine, we'll
signal we've lost the VM in all circumstances that way.

Backported from cc-oci-runtime:

 https://github.com/01org/cc-oci-runtime/commit/2b585d9c8a2b1295479faecddea2cd54df1c3a7f

Fixes: https://github.com/01org/cc-oci-runtime/issues/622
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>