fdleak: Don't let Compare() return true when it shouldn't
When we've already failed the comparison for a previous fd, nothing
should be able to make us return true.

Unfortunately, we were overriding unconditionally the equal boolean
with no regard of the past.

This was found testing the proxy on go tip (what will become 1.9), the
runtime has changed and exposed that bug.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>