vm: Don't fill process.Stderr when asking for a pty
When asking for a pty, we really have only one two streams to the VM
process, one for data going in (stdin) one for data going out
(stdout/err). We asking for a terminal after all.

In that case, we don't want to allocate a sequence number to
process.Stderr and hyperstart will actually reject the command if we do.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>