vm_test: Make sure to stop the mock before to close VM sockets
Because of the recent changes from the proxy combined with the ones
from virtcontainers library, we have to stop the hyperstart/VM mock
before to try to close the communication sockets previously opened.

Indeed, in order to prevent from race condition when closing sockets,
we make sure the connection is ended from the other end of the pipe
instead of forcing on our end.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>