proxy: Make newcontainer/execcmd wait until a shim is ready
I don't really want to have to queue stdout/err data in the proxy.

To solve that, let's introduce a new constraint. The data path:

  shim <-> proxy <-> hyperstart

has to be fully setup before we allow newcontainer/execmd to execute a
new process. While we advise to start the shim as early as possible and
issue the newcontainer/execcmd command after that, we need to
synchronize somewhere. The easier way of doing that is to stall the
newcontainer/execcmd commands until we see the corresponding shim
registering itself.

Fixes: https://github.com/clearcontainers/proxy/issues/21
Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>