proxy: Reset the token status when a shim quits
We close the shim connection when something bad happens:
  - We receive an error trying to write to the socket (most likely
  because the shim died or exited).
  - We have an other kind of unrecoverable error for that client (we
  don't have one right now but we will in the near future)

We now want to progress a bit on the recovery side of things. If the
shim dies, we want to allow it to reconnect and re-claim the session.
This commit does just that.

This is tested by a subsequent unit test: TestShimSendStdinAfterExeccmd

Updates: https://github.com/clearcontainers/proxy/issues/4

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>