proxy: Initialize a client to be runtime by default
A client can be either a shim or a runtime, so change the bit field to
be an enum.

Maybe we can get rid of that at a later point. Having a session
associated with a client is the same thing as being a shim.

This fixes warnings from deadcode and varcheck.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>