vm: Assign a container ID to each session
In case the shim receives a signal, it sends it to the proxy and the
proxy has to figure out which container is affected by this signal.
The current implementation provides the VM ID (called also pod ID),
but this is not right in case we have the pod ID different from the
container ID, or in case several containers are running on the same
pod.

This patch creates a new field containerID in the Session structure
so that we can fill this container ID when the command newcontainer
is issued and actually use this information when a signal has to be
issued.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>