build: Unbreak DefaultSocketPath.
PR #87 inadvertently broke the build setting the default socket path
variable. This appears to be due to a bug in "go build" whereby it
accepts any number of -ldflags "-X ..." options, but only actually
honours the last one. On PR #87 this meant that the Version variable
was set, whilst DefaultSocketPath was silently ignored and its value
defaulted to "".

Fixes #90.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>