proxy: Handle KILL and TERM signals when process not running
Because a state transition from "ready" to "stopped" is possible
for a container created by virtcontainers/runtime, we can end up
with situations where the proxy waits for the shim process to start
so that it can forward a KILL or TERM signal, leading to a deadlock.
Instead, we want the proxy to be smarter, meaning terminate the shim
by himself. Indeed, there is no process/container running inside the
VM, meaning that the corresponding shim will never receive the exit
code from the VM unless the proxy takes care of this.

Fixes #97

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>