proxy: Remove tokenToSession map from proxy.
tokenToSession map is used to easily find a session associated
with a token.

The a session reference was keep in muitples places:

- proxy.tokenToSession[token]
- proxy.vms[container-id].tokenToSession[token]
- proxy.vms[container-id].ioSessions[sequence-number]

Removing tokenToSession the references are keep only in vm struct.

Also, disconnectShim was removing sessions from vm struct but not from
this map leading to never free tokens added to tokenToSession.

Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>