CI: remove tests from run.sh script
The run.sh script has docker tests that are already
covered by our testsuite in the tests repository.
This commit removes those tests.

This commit also changes the way the tests are called in
the tests repository to allow execute OpenShift tests if
the OPENSHIFT_CI envar is set to true.

Fixes: #139

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>