stdin: Forward all stream input coming from the shim
In case the proxy was getting an empty message through the stdin
stream from the shim, it was not sending anything to the agent.
Unfortunately, doing such a thing was preventing the agent from
receiving EOF on stdin.

This patch removes this special handling case, and forward every
input from the shim to the agent.

Fixes #169

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>